feat(og-image): new option to disable existing meta tag overriding, completed some of the ut
