+++ packages/vitepress-plugin-og-image/src/vitepress/constants.ts
@@ -1,0 +1,3 @@
+import { cyan, gray } from 'colorette'
+
+export const logModulePrefix = `${cyan(`@nolebase/vitepress-plugin-og-image`)}${gray(':')}`
